requirements.txt
python-telegram-bot==20.7
gspread==6.0.0
google-auth==2.23.4
google-auth-oauthlib==1.2.4
python-dateutil==2.9.0.post0
pytz==2024.1